+++ app/src/main/java/ru/gb/appgitusers/App.kt
@@ -3,9 +3,8 @@ package ru.gb.appgitusers
 import android.app.Application
 import android.content.Context
 import androidx.fragment.app.Fragment
-import ru.gb.appgitusers.domain.APIGitUserRepository
+import ru.gb.appgitusers.data.APIGitUserRepository
 import ru.gb.appgitusers.domain.IGitUserRepository
-import ru.gb.appgitusers.domain.LocalGitUserRepository
 
 class App : Application() {
     val userRepo: IGitUserRepository by lazy {
