+++ app/src/main/java/ru/gb/appgitusers/domain/GitUserEntity.kt
@@ -1,7 +1,10 @@
 package ru.gb.appgitusers.domain
 
+import com.google.gson.annotations.SerializedName
+
 data class GitUserEntity(
     val login: String,
     val id: String,
-    val url: String
+    @SerializedName("avatar_url")
+    val avatarUrl: String
 )
